+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/adapter/HandlerAdapterNotFoundException.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.adapter;
 
 public class HandlerAdapterNotFoundException extends RuntimeException {
