+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/adapter/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.adapter;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
