+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/adapter/impl/AnnotationHandlerAdapter.java
@@ -1,6 +1,7 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.adapter.impl;
 
 import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.mapping.impl.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.mapping.impl.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
